Move `cfpb-utilities` to `utilities`
