+++ packages/cfpb-design-system/src/elements/base/base.scss
@@ -1,6 +1,6 @@
 @use 'sass:math';
 @use '../abstracts' as *;
-@use '../cfpb-utilities/skeleton' as *;
+@use '../utilities/skeleton' as *;
 @use '../cfpb-tagline/skeleton' as *;
 @use '../cfpb-form-search/skeleton' as *;
