+++ packages/cfpb-design-system/src/elements/cfpb-expandable/index.js
@@ -1,5 +1,5 @@
 import { LitElement, html, css, unsafeCSS } from 'lit';
-import { defineComponent } from '../cfpb-utilities/shared-config';
+import { defineComponent } from '../utilities/shared-config';
 import styles from './styles.component.scss?inline';
 import { CfpbIcon } from '../cfpb-icon';
 import { MaxHeightTransition } from '../../utilities/transition/max-height-transition';
