+++ packages/cfpb-design-system/src/elements/cfpb-expandable/styles.component.scss
@@ -2,7 +2,7 @@
 @use '@cfpb/cfpb-design-system/src/elements/abstracts' as *;
 @use '@cfpb/cfpb-design-system/src/elements/base' as *;
 @use '@cfpb/cfpb-design-system/src/utilities' as *;
-@use '../cfpb-utilities/transition/transition.css' as *;
+@use '../utilities/transition/transition.css' as *;
 
 :host {
   // Theme
