+++ packages/cfpb-design-system/src/elements/cfpb-file-upload/index.js
@@ -1,5 +1,5 @@
 import { LitElement, html, css, unsafeCSS } from 'lit';
-import { defineComponent } from '../cfpb-utilities/shared-config';
+import { defineComponent } from '../utilities/shared-config';
 import { ref, createRef } from 'lit/directives/ref.js';
 import styles from './styles.component.css?inline';
 import { CfpbButton } from '../cfpb-button';
