+++ packages/cfpb-design-system/src/elements/cfpb-flag-usa/index.js
@@ -1,5 +1,5 @@
 import { LitElement, css, unsafeCSS } from 'lit';
-import { defineComponent } from '../cfpb-utilities/shared-config';
+import { defineComponent } from '../utilities/shared-config';
 import styles from './styles.component.css?inline';
 
 /**
