+++ packages/cfpb-design-system/src/elements/cfpb-icon/index.js
@@ -1,8 +1,5 @@
 import { html, LitElement, css, unsafeCSS } from 'lit';
-import {
-  defineComponent,
-  getSharedConfig,
-} from '../cfpb-utilities/shared-config';
+import { defineComponent, getSharedConfig } from '../utilities/shared-config';
 import styles from './styles.component.css?inline';
 
 /**
