+++ packages/cfpb-design-system/src/elements/cfpb-pagination/index.js
@@ -1,8 +1,8 @@
 import { LitElement, html, css, unsafeCSS } from 'lit';
-import { defineComponent } from '../cfpb-utilities/shared-config';
+import { defineComponent } from '../utilities/shared-config';
 import styles from './styles.component.scss?inline';
 import { CfpbIcon } from '../cfpb-icon';
-import { I18nService, MediaQueryService } from '../cfpb-utilities/';
+import { I18nService, MediaQueryService } from '../utilities/';
 
 /**
  *
