+++ packages/cfpb-design-system/src/elements/utilities/i18n-service.js
