+++ packages/cfpb-design-system/src/elements/utilities/parse-child-data.js
