+++ packages/cfpb-design-system/src/elements/utilities/search-service.spec.js
