+++ packages/cfpb-design-system/src/elements/utilities/shared-config.js
