+++ packages/cfpb-design-system/src/elements/utilities/skeleton.css
